backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.3.0
httpx==0.25.2

# File processing
python-magic==0.4.27
pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.8
weasyprint==60.1

# Email
aiosmtplib==3.0.1
